fix(ui): resolve hero detail review findings
